Corrected the declaration of various GAP representations
Correct various issues with the declarations of the representations for
numerical and affine semigroups as well as their ideals:
* they are *not* additive or multiplicative elements in the sense of GAP
* IsAttributeStoringRep implies IsComponentObjectRep, thus there is no
  need to specify the latter
* None of the three representations ever set a component or attribute
  named "list", so don't claim so in the declaration
* for ideals, however, UnderlyingNSIdeal and GeneratorsIdealNS are
  always set, so declare that

Also use ObjectifyWithAttributes when creating ideals, for a tiny
efficiency improvement.

Finally, change various InstallMethod calls to InstallOtherMethod
(this is necessary because due to the above declaration fixes, numerical
semigroups etc. no longer are marked as being ring elements).